Revert "Disable 055-enum-performance on tradefed"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f950901f-b946-4290-9449-6239d5e59674). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icd1dc0b21411a2af6462d1368258444ba7f203da
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com