Revert "Delete SDM and SDC files on a successful dexopt."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b733a98b-81da-4be0-8f73-29d0beba018a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I693d171e8a1961ba4af8539fb4e93d3b8a080fb1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com