Revert "Count SDM files into TYPE_DEXOPT_ARTIFACT in ART-managed..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b733a98b-81da-4be0-8f73-29d0beba018a). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I255ee8bd1d12f8a091fce4a1fee51cd256411377
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com