Simplify oatdump logic for --export-dex-to.
The logic for --export-dex-to was added 8 years ago by
https://r.android.com/505156. It loads the vdex file after the OatFile
object is there, which is unnecessary today because the vdex file must
be loaded during the creating of the OatFile object and can be accessed
through `OatFile::GetVdexFile`. Maybe the OatFile object could be
created without a vdex file 8 years ago, but it's no longer the case.

Moreover, the purpose of it getting the VdexFile object was to call
`VdexFile::Unquicken` to "unquicken" the dex file, but quickening is not
a thing any more. Today, we can simply get the dex file from OatFile
and we don't even need to get the VdexFile object.

Bug: 325430813
Test: m art_oatdump_tests && \
  out/host/linux-x86/testcases/art_oatdump_tests/x86_64/art_oatdump_tests \
  --gtest_filter="*/OatDumpTest.TestExportDex/*"
Test: Call oatdump with --export-dex-to and see the output being
  identical before and after the change.
Change-Id: Ib02263bde962730cb6cae3d488c8339f0dc766e6