Revert "Allow APEX versions in the OAT header to be overriden."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9bb31acb-c0ab-482a-a0b7-186ac5c1a63e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ie1c601e4d031baf865964cbe14b872c1e6f6c525
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com