Revert "Allow APEX versions in the OAT header to be overriden."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/9bb31acb-c0ab-482a-a0b7-186ac5c1a63e). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I8b2b26450dd02aa548790fd2d037933ecca6f0c8
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com