Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/3e93273f-dcca-4417-993d-5ab962c71a61). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6f054cbc1469baaa7c025032078fb0b15276c83a
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com