Revert "Allow APEX versions in the OAT header to be overriden."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/a26285c0-223c-44c8-baa8-790c14a18ff2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ic1895a175cf6d52c54b66103325000d05e168088
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com