Revert "Priority fixes"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/0044566a-51ec-436d-8962-9ae28a3967d4). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 406285562
Change-Id: If3353f460f48b5646e322aaa41b01d3eecb9be1c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com