Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ce5b6502-70b9-4797-a47e-56d6d60ca8a2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib3491c23d97e93b8cfc8d0032ca988ef3dbd105d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com