Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ea62351e-c78e-435b-9497-5ecf0f938318). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Icc01bf13d53e71e16f640fa805a3b4e89d9fdb37
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com