Revert "Enable thread exit checks for bionic"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/81df8da9-3499-49f1-99ec-2c751039b771). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Bug: 406581902
Change-Id: I14a096352b8df86c6bbc8f0ed46b0ad34d081590
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com