Revert "Revert "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/30c34670-8030-416b-81b9-6c96f2515ec2). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib4d8832b13a6823814d82d5879cfb765bd711c74
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com