Revert "Add helper classes for reading and writing SDC files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ada8d8ba-0b7a-46c9-9a8f-978c77c8f926). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0f6d7bd4f2353e7c0ba0747c62b5d3809ed0f0bd
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com