Revert "Allow APEX versions in the OAT header to be overriden."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/ada8d8ba-0b7a-46c9-9a8f-978c77c8f926). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I2065e20364507a6f8d5340bd060f067afc299f57
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com