Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/2f7577ab-bc98-438f-a809-b8e193f713fd). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I492ef110f991b28eedb7ffabc9a8b6b96f82bcad
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com