Revert "Move register check to IsInliningEncouraged"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/f6172291-dcba-4d1e-bd84-efacaf4db6f6). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I00c48d44b492d1d0a99ccda91cdba15894ccf23d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com