Revert "Move register check to IsInliningEncouraged"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/312b3c61-2cbb-43c2-adde-498228cda77d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I0647284eaa6eb1ed0c194dbccdd0c2bf5a86657c
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com