Implement lock dependency tracking in ART
Mutexes should generally be taken in a fixed ordering: if you take
mutex `next` while holding mutex `previous`, you should _never_ hold
`next` when attempting to take `previous`. Otherwise, if another thread
goes through the first codepath, while you're going through the second,
you can end up in a scenario where both threads hold one of the mutexes
and is waiting on the other and deadlock.

This commit implements tracking and enforcement of ordering of Java
synchronized blocks by recording which mutexes are held when a mutex is
taken, and looking for cycles in that graph.

Currently, this only finds lock ordering violations between two nodes,
but I've never actually seen a deadlock caused by three or more mutexes
being taken in different orders, so the task of extending this to full
graph cycle detection is left to the reader.

Change-Id: Ie7c51c950d5b033430917fd512df8fca468bce87