Revert "Reland "Implement if instructions in fast baseline compi..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/4c0a3b49-61d1-423d-84f9-3163ee95330d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ib54d0bb755257617a79a5e74e7c5453fa27baa60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com