Revert "Reland "Pass the dex register array to NterpGetInstanceF..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/167e2922-6fb8-4f2a-a885-c56a72e8b755). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ia53e06177f7c0b5644d03267218bad866a81e7d7
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com