Revert "Delete SDM files on "pm delete-dexopt" and "pm compile -..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Id1a97ca404f8f65e8d1692e80f775c6f26bc48e9
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com