Revert "Update the file GC to clean up SDM files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I006db11a56777101682c0eae3a012611c622fa60
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com