Revert "Delete SDM and SDC files on a successful dexopt."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I98e7a007b93722b36e98460b7f1a8a5ae3a33177
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com