Revert "Delete SDM and SDC files on a successful dexopt."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iae88e5b3b6592c359b81a8af82aa42bd3fefee2d
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com