Revert "Count SDM files into TYPE_DEXOPT_ARTIFACT in ART-managed..."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I6018ecdf83468fbf2c9865e43d37f79a23aae606
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com