Revert "Support loading SDM files at runtime."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/58dfef0a-cdce-4389-8e40-e95fb8d8cbdc). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I52cfefabdcd6ea8efc65030b584b44d3c2092d68
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com