Revert "Add helper classes for reading and writing SDC files."
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/b2084571-2a6b-4aae-ad17-98272df32465). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: I3eecbfc4938f979bdc53c2637f578368cd5af3b0
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com