Avoid calling pthread_getcpuclockid on dead thread
It is possible for an app:
- to spawn a native thread via JNI (e.g. via std::thread)
- this thread can then attach to the runtime (via AttachCurrentThread)
- and exit without calling DetachCurrentThread.

This leaves the runtime's list of threads and the actual threads in an
inconsistent state but does not immediately cause a crash. One thing
that can later trigger a crash once the app is in this inconsistent
state is profiling via `Debug.startMethodTracingSampling`. As part of
profiling the profiler thread call `GetSample` which ends up calling
`pthread_getcpuclockid` with the cached pthread - which is now invalid.
- causing a crash.

In these cases the app is incorrect and should be fixed to call
DetachCurrentThread however:
- Often the user will be reaching for debug tools like
 `Debug.startMethodTracingSampling` precisely to understand a buggy app.
- Library leaking the thread, the library doing the profiling, and
  the app itself - may be implemented by three different companies
  complicating the fix.

This CL adds a check to avoid a subset of these cases cheaply. In a
common case the zombie thread will never have a valid stack so if we
observe:
- there is no previous stack
- the current stack has no frames
we early out and do not record the stack or call pthread_getcpuclockid.

Bug: https://github.com/getsentry/sentry-java/issues/2604
Test: run https://github.com/supervacuus/invalid_pthread_repro

Change-Id: I06bc52ce393721da322377361655dc20d5676f14