Revert "Move register check to IsInliningEncouraged"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/64d35f36-97fc-47e4-84f8-a1464904de2d). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Iabd69c35ca9897b6a98e7e59c4d2ed74427668e1
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com