Revert "Move register check to IsInliningEncouraged"
This is an Android Culprit Assistant revert, created as part of a culprit search session (http://go/aca-get/76da126b-be4c-4647-8f4c-e6bbb7c4afce). It will not be committed, and can be ignored. If you need to commit this revert, please don't use this - create a new revert of the original topic.

Change-Id: Ife1027fae6d6f67372be0afb3f9ecfe65bc946c3
Signed-off-by: boq-android-culprit-assistant@system.gserviceaccount.com