Sync parent process before callbacks execution
This change fixes lack of synchronization between parent and child
processes when a new process group is used for the child process. Lack
of synchronization leads to a cancellation callback silent failure
because callback's kill is performed on the whole process group
`kill(-pid, SIGKILL)` which may still not be set by the child process at
the callback execution time. Thus CancellationSignal.cancel() can skip
process kill due to kill failure. This could be observed when cancel()
is called for package dexopting during the profman execution.

Details:
 * Add the 'cancelRightBeforeForkTest' test to ArtdTest;
 * Prepare a new helper class in runtime/exec_utils.cc to provide
   process synchronization;
 * Update ExecWithoutWait signature.

Test: atest ArtGtestsTargetInstallApex -- --module \
      ArtGtestsTargetInstallApex --test '*cancelRightBeforeForkTest*'

Change-Id: I3bbeda7a83cf50d113bbaf3fc028b2b008bd6598
Signed-off-by: Ivan Titarenko <i.titarenko@samsung.com>