dex2oat: Produce oat and vdex files atomically
Summary: This change prevents partial oat and vdex files being written if dex2oat
dies before finishing. It only affects the path when --oat-file is used,
not the --oat-fd style output. It works by writing to temporary files
and only once the files are flushed and closed, renaming the files to the
final output.

This change was motivated by corrupted oat files we saw in the wild, we
are hoping to eliminate this path as a potential cause.

Note: The ELF writer had to be modified to accept the ELF path separately from
the FILE* because this path is used to write the SONAME property in the
final ELF. In order to preserve the byte-identical output between using
--oat-file vs --oat-fd the final filepath is provided rather than the
.tmp suffixed path. `Dex2oatImageTest:TestExtension` asserts they are
identical.

Test: Build and flash the OS to a device.

Run `atest art_standalone_dex2oat_tests`

I've added new tests to ensure that a killed dex2oat process produces
only a valid oat file or temporary files that can be discarded and that
these temp files do not interfere with subsequent runs and will be
cleaned up correctly.

Change-Id: I89f37e55b5f3db162c6050891ba35494f1671e2e
Signed-off-by: Cian Ruane <cianr@meta.com>